Restrict inline translations to users that can translate.
We can't memoize current_user in I18nline module, we need to "ask" to
the application, otherwise we do not know if the memoized current user
has logged out.

In I18nline::ActionViewExtension we remove the "<span
calss="translation_missing">..." when there is no user or the user does not have translation
privileges.